requirements.txt
playwright
google-generativeai
aiohttp
pydantic
